Added a pretty smart formula
A pretty smart mathematical function that calculates the amount in the last row in a geometric series if you have the sum. Necessary because I needed to know how many nodes there should be on the last step in a tree with a total of 10 nodes. If each row in the tree would been full, the total would be 15.